requirements.txt
streamlit
google-genai
python-dotenv
numpy
pypdf